requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

cffi==1.15.0

cryptography==37.0.2

pillow==9.1.1

pycparser==2.21

